feat/reafctor-core-auth-controllers (#416)
