+++ controllers/coreControllers/authJwtController/isValidAdminToken.js
@@ -1,0 +1,60 @@
+const jwt = require('jsonwebtoken');
+
+const mongoose = require('mongoose');
+
+const Admin = mongoose.model('Admin');
+
+require('dotenv').config({ path: '.variables.env' });
+
+const isValidAdminToken = async (req, res, next) => {
+  try {
+    const token = req.cookies.token;
+
+    if (!token)
+      return res.status(401).json({
+        success: false,
+        result: null,
+        message: 'No authentication token, authorization denied.',
+        jwtExpired: true,
+      });
+
+    const verified = jwt.verify(token, process.env.JWT_SECRET);
+
+    if (!verified)
+      return res.status(401).json({
+        success: false,
+        result: null,
+        message: 'Token verification failed, authorization denied.',
+        jwtExpired: true,
+      });
+
+    const admin = await Admin.findOne({ _id: verified.id, removed: false });
+    if (!admin)
+      return res.status(401).json({
+        success: false,
+        result: null,
+        message: "Admin doens't Exist, authorization denied.",
+        jwtExpired: true,
+      });
+    if (admin.isLoggedIn === 0)
+      return res.status(401).json({
+        success: false,
+        result: null,
+        message: 'Admin is already logout try to login, authorization denied.',
+        jwtExpired: true,
+      });
+    else {
+      req.admin = admin;
+      next();
+    }
+  } catch (err) {
+    res.status(503).json({
+      success: false,
+      result: null,
+      message: err.message,
+      error: err,
+    });
+  }
+};
+
+module.exports = isValidAdminToken;
