+++ controllers/middlewaresControllers/createCRUDController/listAll.js
@@ -8,7 +8,7 @@ const listAll = async (Model, req, res) => {
       .sort({ created: sort })
       .populate();
 
-    if (count > 0) {
+    if (result.length > 0) {
       return res.status(200).json({
         success: true,
         result,
